requirements.txt
tensorflow-gpu==1.7.0
numpy==1.14.2
scipy==0.18.1
ray==0.3.1
